refactored file names -> snake case; added template code for a virtual_bot
