--- basestation/bot/virtualbot/virtualbot.py
@@ -1,2 +1,0 @@
-class VirtualBot(object):
-    pass
